+++ R/run.R
@@ -53,6 +53,11 @@
 #'   of the standard error. A chunk can be as small as a single character.
 #'   At most one of `stderr_line_callback` and `stderr_callback` can be
 #'   non-`NULL`.
+#' @param stderr_to_stdout Whether to redirect the standard error to the
+#'   standard output. Specifying `TRUE` here will keep both in the
+#'   standard output, correctly interleaved. However, it is not possible
+#'   to deduce where pieces of the output were coming from. If this is
+#'   `TRUE`, the standard error callbacks  (if any) are never called.
 #' @param env Environment of the child process, a named character vector.
 #'   IF `NULL`, the environment of the parent is inherited.
 #' @param windows_verbatim_args Whether to omit the escaping of the
@@ -97,7 +102,8 @@ run <- function(
   command = NULL, args = character(), error_on_status = TRUE, wd = NULL,
   echo_cmd = FALSE, echo = FALSE, spinner = FALSE,
   timeout = Inf, stdout_line_callback = NULL, stdout_callback = NULL,
-  stderr_line_callback = NULL, stderr_callback = NULL, env = NULL,
+  stderr_line_callback = NULL, stderr_callback = NULL,
+  stderr_to_stdout = FALSE, env = NULL,
   windows_verbatim_args = FALSE, windows_hide_window = FALSE,
   encoding = "", cleanup_tree = FALSE) {
 
@@ -111,17 +117,19 @@ run <- function(
   assert_that(is.null(stdout_callback) || is.function(stdout_callback))
   assert_that(is.null(stderr_callback) || is.function(stderr_callback))
   assert_that(is_flag(cleanup_tree))
+  assert_that(is_flag(stderr_to_stdout))
   ## The rest is checked by process$new()
   "!DEBUG run() Checked arguments"
 
   if (!interactive()) spinner <- FALSE
 
   ## Run the process
+  stderr <- if (stderr_to_stdout) "2>&1" else "|"
   pr <- process$new(
     command, args, echo_cmd = echo_cmd, wd = wd,
     windows_verbatim_args = windows_verbatim_args,
     windows_hide_window = windows_hide_window,
-    stdout = "|", stderr = "|", env = env, encoding = encoding,
+    stdout = "|", stderr = stderr, env = env, encoding = encoding,
     cleanup_tree = cleanup_tree
   )
   "#!DEBUG run() Started the process: `pr$get_pid()`"
@@ -200,7 +208,7 @@ run_manage <- function(proc, timeout, spinner, stdout_line_callback,
       }
     }
 
-    newerr <- proc$read_error(2000)
+    newerr <- if (proc$has_error_connection()) proc$read_error(2000)
     if (length(newerr) && nzchar(newerr)) {
       stderr <<- paste0(stderr, newerr)
       if (!is.null(stderr_callback)) stderr_callback(newerr, proc)
@@ -262,7 +270,8 @@ run_manage <- function(proc, timeout, spinner, stdout_line_callback,
 
   ## We might still have output
   "!DEBUG run() reading leftover output / error, process `proc$get_pid()`"
-  while (proc$is_incomplete_output() || proc$is_incomplete_error()) {
+  while (proc$is_incomplete_output() ||
+         (proc$has_error_connection() && proc$is_incomplete_error())) {
     proc$poll_io(-1)
     do_output()
   }